ct-master/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
black==21.9b0
bootstrap4==0.1.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.1
confusable-homoglyphs==3.2.0
cycler==0.11.0
Django==3.2.5
django-crispy-forms==1.13.0
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-environ==0.7.0
django-multiselectfield==0.1.12
django-phonenumber-field==6.0.0
django-registration==3.2
django-static-jquery3==5.1.0
django-yearmonth-widget==0.2.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.3.2
fonttools==4.34.4
idna==3.3
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.2
month==1.2.0
mypy-extensions==0.4.3
numpy==1.23.1
packaging==21.3
pandas==1.5.1
pathspec==0.9.0
phonenumbers==8.12.39
Pillow==9.2.0
platformdirs==2.3.0
PyJWT==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tomli==1.2.1
twilio==7.3.2
typing-extensions==3.10.0.2
urllib3==1.26.7
xlwt==1.3.0
